digits: remove duplication
