+++ provekit/common/src/witness/digits.rs
@@ -1,4 +1,9 @@
-use serde::{Deserialize, Serialize};
+use {
+    crate::FieldElement,
+    ark_ff::{BigInt, BitIteratorLE, PrimeField},
+    ark_std::Zero,
+    serde::{Deserialize, Serialize},
+};
 
 /// Allocates witnesses for the digital decomposition of the given witnesses
 /// into its digits in the given bases.  A log base is specified for each digit
@@ -18,4 +23,85 @@ pub struct DigitalDecompositionWitnesses {
     pub first_witness_idx:          usize,
     /// The number of witnesses written to
     pub num_witnesses:              usize,
+}
+
+/// Compute a mixed-base decomposition of a field element into its digits, using
+/// the given log bases. Decomposition is little-endian.
+/// Panics if the value provided can not be represented in the given bases.
+pub fn decompose_into_digits(value: FieldElement, log_bases: &[usize]) -> Vec<FieldElement> {
+    let num_digits = log_bases.len();
+    let mut digits = vec![FieldElement::zero(); num_digits];
+    let mut value_bits = field_to_le_bits(value);
+    let ref mut ref_value_bits = value_bits;
+    // Grab the bits of the element that we need for each digit, and turn them back
+    // into field elements.
+    for (i, log_base) in log_bases.iter().enumerate() {
+        let digit_bits = ref_value_bits.take(*log_base).collect::<Vec<bool>>();
+        digits[i] = le_bits_to_field(&digit_bits);
+    }
+
+    let mut remaining_bits = value_bits;
+    assert!(
+        remaining_bits.all(|bit| !bit),
+        "Higher order bits are not zero"
+    );
+    digits
+}
+
+/// Decomposes a field element into its bits, in little-endian order.
+/// Probably a lot of allocation because it goes to bits, but for little-endian
+/// and big endian is on byte level not on bit level
+pub fn field_to_le_bits(value: FieldElement) -> BitIteratorLE<BigInt<4>> {
+    BitIteratorLE::new(value.into_bigint())
+}
+
+/// Given the binary representation of a field element in little-endian order,
+/// convert it to a field element. The input is padded to the next multiple of 8
+/// bits.
+pub fn le_bits_to_field(bits: &[bool]) -> FieldElement {
+    let le_byte_vec: Vec<u8> = bits
+        .chunks(8)
+        .map(|chunk_in_bits| {
+            chunk_in_bits
+                .iter()
+                .enumerate()
+                .fold(0u8, |acc, (i, bit)| acc | ((*bit as u8) << i))
+        })
+        .collect();
+    FieldElement::from_le_bytes_mod_order(&le_byte_vec)
+}
+
+#[cfg(test)]
+#[test]
+fn test_decompose_into_digits() {
+    let value = FieldElement::from(3 + 2u32 * 256 + 256 * 256);
+    let log_bases = vec![8, 8, 4];
+    let digits = decompose_into_digits(value, &log_bases);
+    assert_eq!(digits.len(), log_bases.len());
+    assert_eq!(digits[0], FieldElement::from(3u32));
+    assert_eq!(digits[1], FieldElement::from(2u32));
+    assert_eq!(digits[2], FieldElement::from(1u32));
+}
+
+#[cfg(test)]
+#[test]
+// Changing from FieldElement :: from_be to from_le didn't negate this test. So
+// this needs to be extended
+fn test_field_to_le_bits() {
+    let value = FieldElement::from(5u32);
+    let bits: Vec<bool> = field_to_le_bits(value).collect();
+    assert_eq!(bits.len(), 256);
+    assert!(bits[0]);
+    assert!(!bits[1]);
+    assert!(bits[2]);
+    assert!(!bits[254]);
+    assert!(!bits[255]);
+}
+
+#[cfg(test)]
+#[test]
+fn test_le_bits_to_field() {
+    let bits = vec![true, false, true, false, false];
+    let value = le_bits_to_field(&bits);
+    assert_eq!(value.into_bigint().0[0], 5);
 }
